// apps/web/src/lib/api/generated/model/brandingUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { BrandingUpdateFontPair } from './brandingUpdateFontPair';
import type { PaymentDefaults } from './paymentDefaults';

export interface BrandingUpdate {
  accent_color?: string | null;
  font_pair?: BrandingUpdateFontPair;
  footer_text?: string | null;
  logo_document_id?: string | null;
  payment_instructions?: PaymentDefaults | null;
  primary_color?: string | null;
  templates?: Partial<Record<'quote' | 'invoice' | 'receipt' | 'credit_note', string>> | null;
}