// apps/web/src/lib/api/generated/model/templateOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateOutTier } from './templateOutTier';

export interface TemplateOut {
  description: string;
  doc_types: string[];
  font_pair: string;
  key: string;
  name: string;
  paper: string;
  tier: TemplateOutTier;
  version: number;
}